# src/main/resources/public/docs/openapi.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Pulmonis API
...
paths:
...
  /user/signup:
    post:
      tags:
        - User
      summary: Signup
      description: Add a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Person's email which is email
                  example: testing@ut.ee
                password:
                  type: string
                  description: User's password
                  example: testing123
                firstName:
                  type: string
                  description: User's first name
                  example: tes
                lastName:
                  type: string
                  description: User's last name
                  example: ting

      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/login:
    post:
      tags:
        - User
      summary: Log in using an email and a password.
      description: Log in using an email and a password.
      requestBody:
        required: true
        content:
          application/json:
            schema: # Request body contents
              type: object
              properties:
                email:
                  type: string
                  description: Person's email which is email
                  example: testing@ut.ee
                password:
                  type: string
                  description: User's password
                  example: testing123
              required:
                - email
                - password
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /refresh-token:
    post:
      tags:
        - Refresh Token
      summary: Refresh JWT token
      description: Refresh a not expired JWT token.
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/logout:
    delete:
      tags:
        - User
      summary: An endpoint handling the logout operation.
      description: "When a logout was successful, system responts with a status of 200 and a corresponding response body
      (refer to documentation below). If the user is not logged in, currently system responds with a status 401. When an
      unexpected error occurs, system returns status 500."
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataOk"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/update:
    patch:
      tags:
        - User
      summary: Update users.
      description: Update user except for email and a password.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserUpdateResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/change-password:
    put:
      tags:
        - User
      summary: Change user's password
      description: Change user's password. User ID is taken from JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataOk"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/validate-password-token:
    get:
      tags:
        - User
      summary: Validate password reset token
      description: Validate password reset token.
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            description: Password reset token represented as UUID.
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/isValidToken"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/reset-forgotten-password:
    post:
      tags:
        - User
      summary: User's password reset
      description: User's password reset based on password reset token.
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            description: Password reset token represented as UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgottenPasswordChangeRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/isValidToken"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /user/reset-password:
    post:
      tags:
        - User
      summary: User's password reset request
      description: Endpoint for sending user's password recovery mail with reset link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataOk"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PasswordChangeRequest:
      type: object
      properties:
        oldPassword:
          type: string
          example: fingersFingers
          description: Old user's password.
        newPassword:
          type: string
          minLength: 8
          example: loremIpsum
          description: New user's password.
      required:
        - oldPassword
        - newPassword
    ForgottenPasswordChangeRequest:
      type: object
      properties:
        newPassword:
          type: string
          description: User's new password
          example: lorem ipsum
      required:
        - newPassword
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          example: john.doe@gmail.com
          description: User's email.
      required:
        - email
    UserRequest:
      type: object
      properties:
        firstName:
          type: string
          description: Person's firstname
          example: Vasja
        lastName:
          type: string
          description: User's lastname
          example: Pupkin
    SignupResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            email:
              type: string
              description: Person's email which is email
              example: testing@ut.ee
            status:
              type: string
              description: User's status
              example: pending
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            token:
              description: Generated JWT
              type: string
              example: string
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    UserUpdateResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            email:
              description: Person's email which is email
              type: string
              example: testing@ut.ee
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
...
    dataOk:
      type: object
      properties:
        data:
          type: object
          properties:
            ok:
              type: boolean
              description: Successful response.
              example: true
        errors:
          type: array
          nullable: true
          description: List of errors returned by API.
          example: null
        status:
          type: integer
          description: Successful response from API.
          example: 200
    isValidToken:
      type: object
      properties:
        data:
          type: object
          properties:
            isValid:
              type: boolean
              description: Whether password reset token is valid or not.
              example: true
        errors:
          type: array
          nullable: true
          description: List of errors returned by API.
          example: null
        status:
          type: integer
          description: Successful response from API.
          example: 200
    EmptyError:
      type: array
      nullable: true
      description: List of errors returned by API.
      example: null
    SuccessfulStatus:
      type: integer
      example: 200
      default: 200
      description: Successful response from API.
    UnsuccessfulStatus:
      type: integer
      example: 400
      description: Successful status response from API. Can vary from 400 to 500.
    UnsuccessfulResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
        status:
          $ref: "#/components/schemas/UnsuccessfulStatus"
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: Error message returned from API.
          nullable: true
          example: "string"
        code:
          type: string
          description: Error code returned from API.
          example: "string"
security:
  - bearerAuth: [ ]